+++ tree_ocp_qp/x_tree_ocp_qp.c
@@ -238,7 +238,7 @@ void TREE_OCP_QP_CREATE(struct TREE_OCP_QP_DIM *dim, struct TREE_OCP_QP *qp, voi
 
 
 	// align to typical cache line size
-	long long l_ptr = (long long) i_ptr;
+	size_t l_ptr = (size_t) i_ptr;
 	l_ptr = (l_ptr+63)/64*64;
